Add files via upload
The American Energy Market Regulator (AEMR) is responsible for looking after the
United States of America’s domestic energy network. The regulator’s responsibility is to
ensure that America’s energy network remains reliable with minimal disruptions, which
are known as outages. There are four key types of outages:
● Consequential – This outage is caused by an exogenous event or a
consequence of work unrelated to the energy provider (i.e. A transmission line
may be down, which means that a particular power provider may be unable to
send energy to the network)
● Forced – This outage is caused by a situation that has forced the power
generating asset to be unavailable. In other words, it has caught the AEMR by
surprise and was not planned. (i.e. A cooling tower that supplies water to cool
down a power-generating asset has broken down, which has now forced the
power-generating asset offline as the temperature can no longer be regulated
properly.)
● Opportunistic – This outage arises when an energy provider wishes to be
proactive with the maintenance of their assets and believes that it can complete
maintenance on its plant within a 48-hour window. If it takes longer than this, the
outage is considered a PLANNED outage as it was planned in advance and will
take longer than 48 hours.
● Planned – This outage arises when an energy provider reports to the AEMR that
an energy-generating asset needs to be taken offline for routine or planned
maintenance activities to ensure the reliability of the asset in the future.
Of the four outage types, the only one the AEMR penalizes is a forced outage, as this
means that, if enough energy providers enter the forced outage state and the demand
for energy is greater than the amount of energy that can be supplied, stress will be put
on the energy system. This will then threaten the reliability of the network which is what
the AEMR wishes to avoid.
Recently, the AEMR management team has been increasingly aware of a large number
of energy providers that submitted outages over the 2016 and 2017 calendar years. The
management team has expressed a desire to have the following two areas of concern
addressed:
A) Energy Stability and Market Outages
B) Energy Losses and Market Reliability